pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "dmemorymunch-mpc"
version = "0.1.0"
description = "Indexed markdown memory MCP server"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "MemoryMunch" }]
dependencies = [
  "mcp>=1.0.0",
  "typer>=0.12.3",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
]

[project.scripts]
dmemorymunch-mpc = "dmemorymunch_mpc.server:main"
dmemorymunch-mpc-admin = "dmemorymunch_mpc.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/dmemorymunch_mpc"]
